add parameter to spe fit
This commit adds a constant offset to the SPE fit. This is to try and
fix the fact that the fit doesn't currently model the space between the
peaks very well. I suspect this has to do with the fact that we are not
modelling the baseline subtraction properly (for example, when there is
a single PE in the region where we subtract the baseline).

Until we can properly model that this constant term helps the fit fit
these regions.